Upgrades to dao and systems
DataAccessObject for save all the users that have been found ever, also
knows if its a safe user and if its a friend, all other systems have
been upgraded to accomodate for this change.

This change was prompted by the fact that the previous system was
sending a matric FUCK ton of gcm messages to other users to check if
they are safe users which started killing the server;'s ability to send
any kind of gcm message.